impl right interface of remove method for all resolvers.

+++ Imagine/Cache/Resolver/AmazonS3Resolver.php
@@ -101,8 +101,12 @@ class AmazonS3Resolver implements ResolverInterface
     /**
      * {@inheritDoc}
      */
-    public function remove($filter, $path = null)
+    public function remove(array $paths, array $filters)
     {
+        if (empty($paths) && empty($filters)) {
+            return;
+        }
+
         if (null === $path) {
             if (!$this->storage->delete_all_objects($this->bucket, sprintf('/%s/i', $filter))) {
                 $this->logError('The objects could not be deleted from Amazon S3.', array(
